[docs] Revisit the docs pages on clearing bundler caches + RN version mismatches
Why
---
The information on the pages to clear the bundler caches was outdated or incorrect in some places. I also did an editorial pass over the writing and formatting to explain some of the motivation behind clearing the caches and to make the commands easier to read.

How
---
For the cache-clearing pages, I researched which files and directories are created by Expo CLI and Metro today. Both `expo start` and `react-native init X; cd X; yarn start` produced the same cache files: `haste-map-XXX` and a directory named just `metro-cache` both under the OS's temp directory.

Additionally, I read through parts of the metro-config and metro-cache source code to be more confident that `metro-cache` is the directory that Metro creates and that it's unlikely there are other cache directories. I also read the parts of the expo-cli and xdl source related to the `--clear` option to understand and explain its behavior with a little more detail (it also tells Webpack to clear its caches).

Other changes:

* Specified "bundler caches" in the titles to be more precise about what kinds of caches we're clearing (we're not clearing any device caches).
* I changed the one-line commands to multi-line blocks that are easier to read and still copy-pasteable.
* `expo start -c` -> `expo start --clear` (in scripts, docs, and other places outside of a terminal, writing out the long-form option helps communicate what's happening better than the shorthand options do)
* Technical corrections (e.g., `watchman watch-del-all` does not clear dependencies)
* OSX -> macOS
* NPM -> npm
* "bare workflow app" -> "bare workflow project" on the version mismatch page (we say "project" everywhere else on that page)

Test Plan
---------
Visited http://localhost:3002/faq/clear-cache-macos-windows/, http://localhost:3002/faq/clear-cache-macos-linux/, and http://localhost:3002/faq/react-native-version-mismatch/.

Verified they all render as expected.